requirements.txt
selene>=2.0.0b4
pytest
attrs==21.4.0
Faker==8.12.1